+++ src/main/java/br/edu/ufersa/multcare/persistence/entities/Medicamento.java
@@ -17,7 +17,7 @@ public class Medicamento implements IEntity {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private long id;
 	private String nome;
 	private String hora;
 	private Date dataInicial;
@@ -36,7 +36,7 @@ public class Medicamento implements IEntity {
 	/* GETTERS */
 
 	public int getId() {
-		return id;
+		return (int) id;
 	}
 	public String getNome() {
 		return nome;
